+++ src/states/shopping-state.h
@@ -39,7 +39,7 @@ class ShoppingState : public GameState
 
     bool CanBeBought(ItemComponent* ic);
 
-    int GetItemStringTotalLen(std::vector<std::unique_ptr<GameObject>>& container);
+    size_t GetItemStringTotalLen(std::vector<std::unique_ptr<GameObject>>& container);
 
     int GetCost(ItemComponent* ic, bool playerSide);
