+++ src/states/target-state.h
@@ -53,7 +53,7 @@ class TargetState : public GameState
 
     std::vector<GameObject*> _targets;
 
-    int _lastTargetIndex = -1;
+    size_t _lastTargetIndex = -1;
     int _maxThrowingRange = 1;
 
     bool _drawHint = false;
